Добавляет  исправления модудяции
