--- js/create-id-generator.js
@@ -1,6 +1,0 @@
-const createIdGenerator = () => {
-  let lastGeneratedId = 0;
-  return () => (lastGeneratedId += 1);
-};
-
-export { createIdGenerator };
