+++ js/data-generation-constants.js
@@ -1,0 +1,17 @@
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const AVATAR_MIN_INDEX = 1;
+const AVATAR_MAX_INDEX = 6;
+const MIN_NUMBER_COMMENTS = 2;
+const MAX_NUMBER_COMMENTS = 10;
+const MAX_GENERATED_OBJECTS = 25;
+
+export {
+  MIN_LIKES,
+  MAX_LIKES,
+  AVATAR_MIN_INDEX,
+  AVATAR_MAX_INDEX,
+  MIN_NUMBER_COMMENTS,
+  MAX_NUMBER_COMMENTS,
+  MAX_GENERATED_OBJECTS
+};
