+++ js/data.js
@@ -1,5 +1,4 @@
-import { createIdGenerator } from './create-id-generator.js';
-import { getRandomInteger, getRandomArrayElement } from './random-integer.js';
+import { createPhotoId, createPhotoUrl, createCommentId, getRandomInteger, getRandomArrayElement, MIN_LIKES, MAX_LIKES, AVATAR_MIN_INDEX, AVATAR_MAX_INDEX, MIN_NUMBER_COMMENTS, MAX_NUMBER_COMMENTS, MAX_GENERATED_OBJECTS } from './main.js';
 
 const DESCRIPTION_PHOTO = ['Вид из окна гостиницы', 'На пляж', 'Красивый пляж', 'Я фотограф', 'А это точно джакузи?', 'Точно дверь не сломалась?', 'Обед, когда до лета 7 дней', 'Очень молодое вино - это сок', 'Привет - железная птица', 'Удобная полка', 'Дорога в рай', 'Ауди', 'Местная кухня', 'Коксоролл', 'Когда Железному человеку холодно', 'Страшно красиво', 'Я первый слева в третьем ряду, точнее мое ухо', 'Американская  мечта', 'Заказал на Кекс-экспресс', 'Какой отдых без пальм?', 'Вкусняшка', 'Попросила мужа красиво сфотографировать на фоне заката)', 'Мой сосед, а может и обед', 'Концерт', 'Похоже повернул не туда'];
 
@@ -13,18 +12,6 @@ const ARRAY_MESSAGES = [
 ];
 
 const USER_NAME = ['Михаил', 'Александр', 'Максим', 'Марк', 'Артём', 'Лев', 'Матвей', 'Иван', 'Дмитрий', 'Тимофей', 'София', 'Анна', 'Мария', 'Ева', 'Виктория', 'Полина', 'Варвара', 'Алиса', 'Василиса', 'Александра'];
-
-const MIN_LIKES = 15;
-const MAX_LIKES = 200;
-const AVATAR_MIN_INDEX = 1;
-const AVATAR_MAX_INDEX = 1;
-const MIN_NUMBER_COMMENTS = 2;
-const MAX_NUMBER_COMMENTS = 10;
-const MAX_GENERATED_OBJECTS = 25;
-
-const createPhotoId = createIdGenerator();
-const createPhotoUrl = createIdGenerator();
-const createCommentId = createIdGenerator();
 
 const createComment = () => ({
   id: createCommentId(),
