+++ js/full-size-mode.js
@@ -1,0 +1,3 @@
+const bigPicture = document.querySelector('.big-picture');
+
+//bigPicture.classList.remove('hidden');
